DEV-270 merge fixes
