+++ src/Multifactor.Radius.Adapter.v2/Core/FirstFactor/LdapAuth/PapAuthProcessor.cs
@@ -3,9 +3,11 @@ using Microsoft.Extensions.Logging;
 using Multifactor.Core.Ldap;
 using Multifactor.Core.Ldap.Connection;
 using Multifactor.Radius.Adapter.v2.Core.Configuration.Client;
+using Multifactor.Radius.Adapter.v2.Core.FirstFactor.BindNameFormat;
 using Multifactor.Radius.Adapter.v2.Core.Ldap;
 using Multifactor.Radius.Adapter.v2.Core.Radius;
 using Multifactor.Radius.Adapter.v2.Infrastructure.Pipeline.Context;
+using ILdapConnection = Multifactor.Radius.Adapter.v2.Core.Ldap.ILdapConnection;
 using ILdapConnectionFactory = Multifactor.Radius.Adapter.v2.Core.Ldap.ILdapConnectionFactory;
 
 namespace Multifactor.Radius.Adapter.v2.Core.FirstFactor.LdapAuth;
@@ -14,12 +16,14 @@ public class PapAuthProcessor : ILdapAuthProcessor
 {
     private readonly ILogger<PapAuthProcessor> _logger;
     private ILdapConnectionFactory _ldapConnectionFactory;
+    private readonly ILdapBindNameFormatterProvider _ldapBindNameFormatterProvider;
 
     public AuthenticationType AuthenticationType => AuthenticationType.PAP;
     
-    public PapAuthProcessor(ILdapConnectionFactory ldapConnectionFactory, ILogger<PapAuthProcessor> logger)
+    public PapAuthProcessor(ILdapConnectionFactory ldapConnectionFactory, ILdapBindNameFormatterProvider ldapBindNameFormatterProvider, ILogger<PapAuthProcessor> logger)
     {
         _ldapConnectionFactory = ldapConnectionFactory;
+        _ldapBindNameFormatterProvider = ldapBindNameFormatterProvider;
         _logger = logger;
     }
     
@@ -58,11 +62,29 @@ public class PapAuthProcessor : ILdapAuthProcessor
         string password)
     {
         var serverConfig = context.LdapServerConfiguration;
+        if (serverConfig is null)
+            throw new InvalidOperationException("No Ldap servers configured.");
+        
+        var bindName = string.Empty;
+        
         try
         {
+            var ldapImpl = context.LdapSchema!.LdapServerImplementation;
+            var formatter = _ldapBindNameFormatterProvider.GetLdapBindNameFormatter(ldapImpl);
+            if (formatter is null)
+                _logger.LogWarning("No LDAP bind name formatter configured for '{implementation}' implementation.", ldapImpl);
+
+            var formatted = string.Empty;
+            if (context.UserLdapProfile is not null)
+                formatted = formatter?.FormatName(login, context.UserLdapProfile);
+            
+            bindName = string.IsNullOrWhiteSpace(formatted) ? login : formatted;
+            
+            _logger.LogDebug("Use '{name}' for LDAP bind.", bindName);
+            
             using var connection = GetConnection(
                 serverConfig.ConnectionString,
-                login,
+                bindName,
                 password,
                 serverConfig.BindTimeoutInSeconds);
 
@@ -72,7 +94,7 @@ public class PapAuthProcessor : ILdapAuthProcessor
         {
             if (ex is not LdapException ldapException)
             {
-                _logger.LogError(ex, "Verification user '{user:l}' at {ldapUri:l} failed", login, serverConfig.ConnectionString);
+                _logger.LogError(ex, "Verification user '{user:l}' at {ldapUri:l} failed", bindName, serverConfig.ConnectionString);
                 return false;
             }
 
@@ -80,14 +102,13 @@ public class PapAuthProcessor : ILdapAuthProcessor
             if (info != null)
                 ProcessErrorReason(info, context, serverConfig);
 
-            _logger.LogWarning(ldapException, "Verification user '{user:l}' at {ldapUri:l} failed: {dataReason:l}", login, serverConfig.ConnectionString, info?.ReasonText);
+            _logger.LogWarning(ldapException, "Verification user '{user:l}' at {ldapUri:l} failed: {dataReason:l}", bindName, serverConfig.ConnectionString, info?.ReasonText);
         }
 
         return false;
     }
     
-    private ILdapConnection GetConnection(string connectionString, string userName, string password,
-        int bindTimeoutInSeconds)
+    private ILdapConnection GetConnection(string connectionString, string userName, string password, int bindTimeoutInSeconds)
     {
         var connectionOptions = new LdapConnectionOptions(
             new LdapConnectionString(connectionString),
