+++ src/Multifactor.Radius.Adapter.v2/Extensions/ServiceCollectionExtensions.cs
@@ -13,6 +13,8 @@ using Multifactor.Radius.Adapter.v2.Core.Configuration.Service;
 using Multifactor.Radius.Adapter.v2.Core.Configuration.Service.Build;
 using Multifactor.Radius.Adapter.v2.Core.FirstFactor;
 using Multifactor.Radius.Adapter.v2.Core.FirstFactor.BindNameFormat;
+using Multifactor.Radius.Adapter.v2.Core.FirstFactor.LdapAuth;
+using Multifactor.Radius.Adapter.v2.Core.FirstFactor.LdapAuth.MsChapV2;
 using Multifactor.Radius.Adapter.v2.Core.Ldap;
 using Multifactor.Radius.Adapter.v2.Core.Radius.Attributes;
 using Multifactor.Radius.Adapter.v2.Infrastructure.Configuration.RadiusAdapter.Build;
@@ -21,6 +23,7 @@ using Multifactor.Radius.Adapter.v2.Infrastructure.Logging;
 using Multifactor.Radius.Adapter.v2.Infrastructure.Pipeline;
 using Multifactor.Radius.Adapter.v2.Infrastructure.Pipeline.Builder;
 using Multifactor.Radius.Adapter.v2.Infrastructure.Pipeline.Steps;
+using Multifactor.Radius.Adapter.v2.Server;
 using Multifactor.Radius.Adapter.v2.Server.Udp;
 using Multifactor.Radius.Adapter.v2.Services.AuthenticatedClientCache;
 using Multifactor.Radius.Adapter.v2.Services.Cache;
